Make Steam Cloud transitions durable and complete

+++ app/src/main/java/app/gamenative/data/UFS.kt
@@ -4,7 +4,7 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class UFS(
-    val quota: Int = 0,
+    val quota: Long = 0L,
     val maxNumFiles: Int = 0,
     val saveFilePatterns: List<SaveFilePattern> = emptyList(),
 )
